+++ domain/src/main/java/com/agileburo/anytype/domain/block/interactor/UpdateTextStyle.kt
@@ -1,31 +1,27 @@
 package com.agileburo.anytype.domain.block.interactor
 
 import com.agileburo.anytype.domain.base.BaseUseCase
-import com.agileburo.anytype.domain.base.Either
 import com.agileburo.anytype.domain.block.model.Block.Content.Text
 import com.agileburo.anytype.domain.block.model.Command
 import com.agileburo.anytype.domain.block.repo.BlockRepository
 import com.agileburo.anytype.domain.common.Id
+import com.agileburo.anytype.domain.event.model.Payload
 
 /**
  * Use-case for udpating a block's text style
  */
 open class UpdateTextStyle(
     private val repo: BlockRepository
-) : BaseUseCase<Unit, UpdateTextStyle.Params>() {
+) : BaseUseCase<Payload, UpdateTextStyle.Params>() {
 
-    override suspend fun run(params: Params) = try {
+    override suspend fun run(params: Params) = safe {
         repo.updateTextStyle(
             command = Command.UpdateStyle(
                 style = params.style,
                 context = params.context,
                 targets = params.targets
             )
-        ).let {
-            Either.Right(it)
-        }
-    } catch (t: Throwable) {
-        Either.Left(t)
+        )
     }
 
     /**